doing naive feature extraction firsh

+++ models/slic_vit.py
@@ -381,8 +381,8 @@ class SLICViT(nn.Module):
         #text_features_normalized = (text_features - torch.min(text_features)) / (torch.max(text_features) - torch.min(text_features))
         #text_features_normalized = torch.tensor(np.load(root_path + "Nesf0_2D/" + text + "_clip_feature.npy"))
         #print(text_features_normalized)
-        text_features_normalized = text_features_normalized.to(torch.float)
-        image_features_normalized = image_features_normalized.to(torch.float)
+        # text_features_normalized = text_features_normalized.to(torch.float)
+        # image_features_normalized = image_features_normalized.to(torch.float)
         for i in range(r):
             for j in range(c):
                 query_map[i,j,0] = torch.dot(image_features_normalized[i,j,:], text_features_normalized) / ((np.linalg.norm(image_features_normalized[i,j,:]) * np.linalg.norm(text_features_normalized)))
